Introduced News

+++ app/src/main/java/it/diegocastagna/ifermi/models/Model.java
@@ -1,18 +1,10 @@
 package it.diegocastagna.ifermi.models;
 
-import android.os.AsyncTask;
-
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
-import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 
 import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Observable;
@@ -29,9 +21,9 @@ import it.diegocastagna.ifermi.utils.RssNews;
  * @since 1.0
  */
 public class Model extends Observable {
-    public final static String rssUrl = "https://www.fermimn.edu.it/?action=rss";
-    public final static String classesUrl = "https://www.fermimn.edu.it/orari/orario_in_corso/";
-    public final static String imageRssUrl = "https://www.fermimn.edu.it/?clean=true&action=icon&newsid=";
+    public final static String RSSURL = "https://www.fermimn.edu.it/?action=rss";
+    public final static String CLASSESURL = "https://www.fermimn.edu.it/orari/orario_in_corso/";
+    public final static String IMAGERSSURL = "https://www.fermimn.edu.it/?clean=true&action=icon&newsid=";
 
     private ArrayList<RssNews> rssNews;
 
@@ -81,30 +73,26 @@ public class Model extends Observable {
         // Checking if Cache file Exists and is not empty
         File f = new File(this.cacheDir, "newsCacheFile.rss");
         if(f.exists()){
-            System.out.println("Esiste");
             try{
                 File f2 = new File(this.cacheDir, "newsCacheFileTemp.rss");
-                //new DownloadFileFromURL().execute(f2, Model.rssUrl);
-                downloadFile(f2, Model.rssUrl);
+                new DownloadFileFromURL().execute(f2, Model.RSSURL).get();
                 // Checking if files are the same TODO: Check Files hash? Ask Salvi
-
-                // If they are different file, it deletes the old one and replace it with the new one
-                if(f.delete() && f2.renameTo(f)){
-                    createRssNews(f);
-                } else {
-                    System.out.println("[MESSAGE] Temporary fallback to Temp File");
-                    createRssNews(f2);
-                }
-
+                if(checkFiles(f,f2)){
+                    // File is not the same it deletes the old one and replace it with the new one
+                    if(f.delete() && f2.renameTo(f)){
+                        createRssNews(f);
+                    } else {
+                        System.out.println("[MESSAGE] Temporary fallback to Temp File");
+                        createRssNews(f2);
+                    }
+                    }
                 f2.delete();
             }catch (Exception e){
                 System.out.println("[ERROR]: During the creation of Rss News Tree in Memory: " + e);
             }
         } else {
             try {
-                //new DownloadFileFromURL().execute(f, Model.rssUrl).get();
-                downloadFile(f, Model.rssUrl);
-                System.out.println("111 Creiamo l'array");
+                new DownloadFileFromURL().execute(f, Model.RSSURL).get();
                 createRssNews(f);
             }catch (Exception e) {
                 System.out.println("[ERROR]: During the creation of Rss News Tree in Memory: " + e);
@@ -113,20 +101,6 @@ public class Model extends Observable {
 
         f.delete();
         return rssNews;
-    }
-
-    private void downloadFile(File f, String urlStr) throws IOException {
-        URL url = new URL(urlStr);
-        InputStream bis = url.openStream();
-        OutputStream fos = new FileOutputStream(f); // Where the data will be saved
-
-        byte[] buffer = new byte[2048];
-        int count;
-        while((count = bis.read(buffer,0,2048)) != -1){
-            fos.write(buffer, 0, count);
-        }
-
-        fos.close();
     }
 
     /**
@@ -141,21 +115,18 @@ public class Model extends Observable {
         Document doc = parser.parse(f); // XML Document
         Element el = doc.getDocumentElement(); // Root Element
         NodeList itemList = el.getElementsByTagName("item"); // List of all Items
-        System.out.println();
         for(int i = 0; i < itemList.getLength(); i++){
             RssNews temp = new RssNews();
-            Node n = itemList.item(i);
-            n.normalize();
-            NodeList newsList = n.getChildNodes();
+            NodeList newsList = itemList.item(i).getChildNodes();
 
-            temp.setTitle(newsList.item(1).getTextContent()); // Get and set the title of the news
-            temp.setDescription(newsList.item(3).getTextContent()); // Get and set the description of the news
-            String tempIconStr = newsList.item(5).getTextContent(); // Full String of the Icon
+            String tempDescription = newsList.item(3).getTextContent().substring(newsList.item(3).getTextContent().lastIndexOf("<br />")+7).trim();
+            String tempIconStr = newsList.item(5).getTextContent().replace("newsid=","news"); // Full String of the Icon
 
-            System.out.println(newsList.item(0).getTextContent()+"1");
-
-            //temp.setIconId(tempIconStr.substring(tempIconStr.lastIndexOf("newsid="))); // Only the Id of the Icon
-            //rssNews.add(temp);
+            temp.setTitle(newsList.item(1).getTextContent().trim()); // Get and set the title of the news
+            temp.setLongDesc(tempDescription); // Get and set the Full description of the news
+            temp.setShortDesc(tempDescription.substring(0,130).trim() + "...");
+            temp.setIconId(tempIconStr.substring(tempIconStr.lastIndexOf("news")+4).trim()); // Only the Id of the Icon
+            rssNews.add(temp);
         }
     }
 
@@ -167,5 +138,9 @@ public class Model extends Observable {
 
         // Checking if Cache file Exists and is not empty
         return null;
+    }
+
+    public boolean checkFiles(File f, File f2){
+        return true;
     }
 }
